refactor: move netplay code from input.go to a new file
